py-model/requirements.txt
spacy==3.7.4
scispacy==0.5.4
negspacy==1.0.4
openai==1.30.1
pydantic==2.7.1
python-dotenv==1.0.1
